refactor to make interface which unify the classes needs to start/stop for app
