+++ platform-control-center/src/main/java/com/flow/platform/cc/context/ContextInitHandler.java
@@ -16,11 +16,7 @@
 
 package com.flow.platform.cc.context;
 
-import com.flow.platform.cc.consumer.QueueConsumer;
-import com.flow.platform.cc.service.ZoneService;
 import com.flow.platform.cc.util.SpringContextUtil;
-import com.flow.platform.cc.util.ZkHelper;
-import com.flow.platform.domain.Zone;
 import com.flow.platform.util.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
@@ -38,31 +34,15 @@ public class ContextInitHandler implements ApplicationListener<ContextRefreshedE
     private final static Logger LOGGER = new Logger(ContextInitHandler.class);
 
     @Autowired
-    protected ZkHelper zkHelper;
-
-    @Autowired
-    private ZoneService zoneService;
-
-    @Autowired
     private SpringContextUtil springContextUtil;
 
     @Override
     public void onApplicationEvent(ContextRefreshedEvent event) {
-        // init root node
-        String path = zoneService.createRoot();
-        LOGGER.trace("Root zookeeper node initialized: %s", path);
-
-        // init zone nodes
-        for (Zone zone : zkHelper.getDefaultZones()) {
-            path = zoneService.createZone(zone);
-            LOGGER.trace("Zone zookeeper node initialized: %s", path);
-        }
-
         // init queue consumer
-        for (String queueConsumeName : springContextUtil.getBeanNameByType(QueueConsumer.class)) {
-            QueueConsumer queueConsumer = (QueueConsumer) springContextUtil.getBean(queueConsumeName);
-            queueConsumer.start();
-            LOGGER.trace("Start queue consumer : %s", queueConsumer.getName());
+        for (String eventClassName : springContextUtil.getBeanNameByType(ContextEvent.class)) {
+            ContextEvent eventClass = (ContextEvent) springContextUtil.getBean(eventClassName);
+            eventClass.start();
+            LOGGER.trace("Start context : %s", eventClassName);
         }
     }
 }
